fix(evm): widen blob_base_fee + blob gas charge to u128
EIP-4844 fixtures probe excess_blob_gas values that drive
fake_exponential outputs above 2^64 (e.g. parent_excess_blobs_1230
expects ~68 bits). Two adjustments:

1. ExecutionContext.blobBaseFee is now __uint128_t. The host writes
   the low 16 bytes of evmc_tx_context.blob_base_fee instead of just
   the low 8, so BLOBBASEFEE returns the right value when blob_base_fee
   exceeds u64.

2. Capa B harness's blob-gas debit was computing
   BLOB_GAS_PER_BLOB * num_blobs * blob_base_fee in u64, which wraps
   for high blob_base_fee (the product easily exceeds 2^64 even when
   blob_base_fee itself fits). The whole multiplication and the
   subsequent sender-balance debit now go through u128 arithmetic
   (new u256SubU128 helper).

Capa B impact: 11607 -> 11613 (+6 fixtures). Small but unblocks the
high-excess decreasing-blob-gas-cost cases that were storing u64-max
instead of the real value.

All 12 EVM Boost suites still green.

+++ src/evm/host.h
@@ -76,8 +76,14 @@ struct ExecutionContext
      *  Computed by the harness/RPC layer from the block header's
      *  excessBlobGas via fake_exponential — at the consensus layer we
      *  don't drive blob-fee market mechanics yet, but the opcode must
-     *  return a sane value for Cancun-fork fixtures and contracts. */
-    uint64_t blobBaseFee{0};
+     *  return a sane value for Cancun-fork fixtures and contracts.
+     *
+     *  Width: __uint128_t — for moderately-elevated excessBlobGas the
+     *  result climbs past u64 (the official fixtures probe this
+     *  range, e.g. parent_excess_blobs_1230 expects a ~68-bit value).
+     *  Stays in the low 16 bytes of the evmc_tx_context.blob_base_fee
+     *  evmc::uint256be field. */
+    __uint128_t blobBaseFee{0};
 };
 
 /**
